Move UI components to a UI directory
