+++ app/src/components/ui/SearchBar.tsx
