+++ app/src/layout/JustBanksy.tsx
@@ -2,10 +2,10 @@ import { albumsOutline, imagesOutline } from 'ionicons/icons';
 import { useState } from 'react';
 import { HStack, VStack } from 'reaction';
 import LoadingAnimation from 'src/components/LoadingAnimation';
-import SearchBar from 'src/components/SearchBar';
-import Sidebar, { SidebarNavigationButton } from 'src/components/Sidebar';
-import background from 'src/resources/images/backgrounds/banksy.png';
-import logo from 'src/resources/images/icons/Just Banksy.png';
+import SearchBar from 'src/components/ui/SearchBar';
+import Sidebar, { SidebarNavigationButton } from 'src/components/ui/Sidebar';
+import background from 'src/assets/images/backgrounds/banksy.png';
+import logo from 'src/assets/images/icons/Just Banksy.png';
 import ArrayUtils from 'src/utils/ArrayUtils';
 import { BanksyKit } from 'src/utils/JustSDK';
 import './JustBanksy.css';
